requirements.txt
altgraph==0.17.4
comtypes==1.4.10
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinsane2==2.0.13
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
